backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic-settings==2.1.0
sqlalchemy==2.0.23
aiosqlite==0.19.0
structlog==23.2.0
slowapi==0.1.4
pydantic==2.5.2
websockets==12.0
apscheduler==3.10.4
pandas==2.1.3
numpy
scikit-learn==1.3.2
xgboost==2.0.2
alembic==1.12.1
asyncpg==0.29.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
shap==0.44.0
scipy
reportlab==4.0.8
python-multipart==0.0.6
redis[hiredis]
httpx
